+++ examples/authors/go/database/sql/models.go
@@ -45,3 +45,15 @@ type UpsertAuthorParams struct {
 	AuthorName string
 	Biography  *string
 }
+
+type FindAuthorsByNameRow struct {
+	ID   uint64
+	Name string
+	Bio  *string
+}
+
+type FindAuthorsByNameCoveringRow struct {
+	ID   uint64
+	Name string
+	Bio  *string
+}